+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- `description()` of `DecodingError` is now returning deprecation message like Rust 1.27.
 
 ## 0.4.0 - 2018-05-23
 ### Added